move the classes to the new architecture
