+++ inc/Engine/Optimization/Buffer/Subscriber.php
@@ -1,16 +1,16 @@
 <?php
-namespace WP_Rocket\Subscriber\Optimization;
+declare(strict_types=1);
 
-use WP_Rocket\Buffer\Optimization;
+namespace WP_Rocket\Engine\Optimization\Buffer;
+
 use WP_Rocket\Event_Management\Subscriber_Interface;
 
 /**
  * Event subscriber to buffer and process a page content.
  *
- * @since  3.3
- * @author Grégory Viguier
+ * @since 3.3
  */
-class Buffer_Subscriber implements Subscriber_Interface {
+class Subscriber implements Subscriber_Interface {
 	/**
 	 * Optimization instance
 	 *
@@ -39,9 +39,7 @@ class Buffer_Subscriber implements Subscriber_Interface {
 	/**
 	 * Start buffering the page content and apply optimizations if we can.
 	 *
-	 * @since  3.3
-	 * @access public
-	 * @author Grégory Viguier
+	 * @since 3.3
 	 */
 	public function start_content_process() {
 		return $this->optimizer->maybe_init_process();
